decodeing the whole circle successfully
one problem is that the predefined array is not accurate.
in the upper circle, the numbers of the block in row 7 is 17, but in the lower circle, the numbers of the block in row 7 is 16. 
so this difference will cause that the parts of text can not be decoded correctly.