fix the bug which can't decode an image
solve the problem which can't decode the image "congrat2D.jpg"
the reason is that this image is upright initially. in previous logic, when i find the center of the circle, what I did next is to find lines in order to get an angle. This is where the problem happen. 
if the image is upright, the HoughLIne() function still can return an angle which is useless and incorrect. if you rotate the circle according the angle. The bug will appear, because the second rotation can not get a correct angle, which I don't know why. it's very wired.

The solution, I changed the logic. At first, I will check whether the image is upright. if so, go to find a pattern directly. if not, use previous logic to process the image.

After this modifying, I can get a pretty good result.