fix the pixel under 1000 problem
in yesterday test, i can not find the circle in the pictures whose pixels are less than 1000.
resize method is not good way, because I can't use the original parameters to get a good circle.
the replacement plan is decreasing the min_radius parameter in HoughCircle as the result that more than 1 circles will be found. but to get the biggest the circle all the time.