organize the code4, getUprightDownAngle()
by reorganizing this function, I found that the reason why the second aligned angle is not correct for some images. 
because I use 20 as offset to locate the pixel in the neighbor block. it's only okay for those pictures in 1000*1000.
so I use the radius / 23 as the offset for instead.